ctx Fixes #81
