+++ src/ast/node/program.rs
@@ -2,12 +2,15 @@ use super::function::FuncDefNode;
 use super::types::StructDefNode;
 use super::*;
 use crate::ast::accumulators::*;
+use crate::ast::builder::llvmbuilder::create_llvm_deps;
 use crate::ast::builder::llvmbuilder::LLVMBuilder;
 use crate::ast::builder::no_op_builder::NoOpBuilder;
 use crate::ast::builder::BuilderEnum;
 use crate::ast::builder::IRBuilder;
 use crate::ast::compiler::{compile_dry_file, ActionType};
-use crate::ast::ctx::{self, create_llvm_deps, Ctx, LSPDef, Mod};
+use crate::ast::ctx::{self, Ctx};
+use crate::ast::plmod::LSPDef;
+use crate::ast::plmod::Mod;
 use crate::lsp::mem_docs::{EmitParams, FileCompileInput, MemDocsInput};
 use crate::lsp::semantic_tokens::SemanticTokensBuilder;
 use crate::lsp::text;
